p4-revocation/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
redis==5.0.8
aiokafka==0.10.0
PyJWT==2.8.0
cryptography==42.0.5
requests==2.31.0
prometheus-client
oqs
pqcrypto
